+++ src/psyclone/transformations.py
@@ -44,7 +44,6 @@
 # pylint: disable=too-many-lines
 
 import abc
-import re
 
 from psyclone import psyGen
 from psyclone.configuration import Config
@@ -58,25 +57,25 @@ from psyclone.psyGen import (Transformation, CodedKern, Kern, InvokeSchedule,
                              BuiltIn)
 from psyclone.psyir.nodes import (
     ACCDataDirective, ACCDirective, ACCEnterDataDirective, ACCKernelsDirective,
-    ACCLoopDirective, ACCParallelDirective, ACCRoutineDirective, Assignment,
-    Call, CodeBlock, Directive, IntrinsicCall, Literal, Loop, Node,
+    ACCLoopDirective, ACCParallelDirective, ACCRoutineDirective,
+    Call, CodeBlock, Directive, Loop, Node,
     OMPDeclareTargetDirective, OMPDirective, OMPMasterDirective,
     OMPParallelDirective, OMPParallelDoDirective, OMPSerialDirective,
     OMPSingleDirective, OMPTaskloopDirective, PSyDataNode, Reference,
-    Return, Routine, Schedule, Statement, WhileLoop)
+    Return, Routine, Schedule)
 from psyclone.psyir.nodes.array_mixin import ArrayMixin
 from psyclone.psyir.nodes.structure_member import StructureMember
 from psyclone.psyir.nodes.structure_reference import StructureReference
 from psyclone.psyir.symbols import (
     ArgumentInterface, DataSymbol, UnresolvedType, INTEGER_TYPE,
-    ScalarType, Symbol, SymbolError, UnsupportedFortranType)
+    ScalarType, Symbol, SymbolError)
 from psyclone.psyir.transformations.loop_trans import LoopTrans
 from psyclone.psyir.transformations.omp_loop_trans import OMPLoopTrans
-from psyclone.psyir.transformations.parallel_loop_trans import \
-    ParallelLoopTrans
+from psyclone.psyir.transformations.parallel_loop_trans import (
+    ParallelLoopTrans)
 from psyclone.psyir.transformations.region_trans import RegionTrans
-from psyclone.psyir.transformations.transformation_error import \
-    TransformationError
+from psyclone.psyir.transformations.transformation_error import (
+    TransformationError)
 
 
 def check_intergrid(node):
@@ -2591,160 +2590,6 @@ class ACCRoutineTrans(Transformation, MarkRoutineForGPUMixin):
         super().validate(node, options)
 
         self.validate_it_can_run_on_gpu(node, options)
-
-
-class ACCKernelsTrans(RegionTrans):
-    '''
-    Enclose a sub-set of nodes from a Schedule within an OpenACC kernels
-    region (i.e. within "!$acc kernels" ... "!$acc end kernels" directives).
-
-    For example:
-
-    >>> from psyclone.parse.algorithm import parse
-    >>> from psyclone.psyGen import PSyFactory
-    >>> api = "nemo"
-    >>> ast, invokeInfo = parse(NEMO_SOURCE_FILE, api=api)
-    >>> psy = PSyFactory(api).create(invokeInfo)
-    >>>
-    >>> from psyclone.transformations import ACCKernelsTrans
-    >>> ktrans = ACCKernelsTrans()
-    >>>
-    >>> schedule = psy.invokes.get('tra_adv').schedule
-    >>> # Uncomment the following line to see a text view of the schedule
-    >>> # print(schedule.view())
-    >>> kernels = schedule.children[9]
-    >>> # Transform the kernel
-    >>> ktrans.apply(kernels)
-
-    '''
-    excluded_node_types = (CodeBlock, Return, PSyDataNode,
-                           psyGen.HaloExchange, WhileLoop)
-
-    def apply(self, node, options=None):
-        '''
-        Enclose the supplied list of PSyIR nodes within an OpenACC
-        Kernels region.
-
-        :param node: a node or list of nodes in the PSyIR to enclose.
-        :type node: (a list of) :py:class:`psyclone.psyir.nodes.Node`
-        :param options: a dictionary with options for transformations.
-        :type options: Optional[Dict[str, Any]]
-        :param bool options["default_present"]: whether or not the kernels \
-            region should have the 'default present' attribute (indicating \
-            that data is already on the accelerator). When using managed \
-            memory this option should be False.
-
-        '''
-        # Ensure we are always working with a list of nodes, even if only
-        # one was supplied via the `node` argument.
-        node_list = self.get_node_list(node)
-
-        self.validate(node_list, options)
-
-        parent = node_list[0].parent
-        start_index = node_list[0].position
-
-        if not options:
-            options = {}
-        default_present = options.get("default_present", False)
-
-        # Create a directive containing the nodes in node_list and insert it.
-        directive = ACCKernelsDirective(
-            parent=parent, children=[node.detach() for node in node_list],
-            default_present=default_present)
-
-        parent.children.insert(start_index, directive)
-
-    def validate(self, nodes, options=None):
-        # pylint: disable=signature-differs
-        '''
-        Check that we can safely enclose the supplied node or list of nodes
-        within OpenACC kernels ... end kernels directives.
-
-        :param nodes: the proposed PSyIR node or nodes to enclose in the
-                      kernels region.
-        :type nodes: (list of) :py:class:`psyclone.psyir.nodes.Node`
-        :param options: a dictionary with options for transformations.
-        :type options: Optional[Dict[str, Any]]
-        :param bool options["disable_loop_check"]: whether to disable the
-            check that the supplied region contains 1 or more loops. Default
-            is False (i.e. the check is enabled).
-
-        :raises NotImplementedError: if the supplied Nodes belong to
-            a GOInvokeSchedule.
-        :raises TransformationError: if there is an access to an assumed-size
-            character variable within the region.
-        :raises TransformationError: if the proposed region contains a call to
-            an intrinsic that is not available on the accelerator.
-        :raises TransformationError: if there are no Loops within the
-            proposed region and options["disable_loop_check"] is not True.
-
-        '''
-        # Ensure we are always working with a list of nodes, even if only
-        # one was supplied via the `nodes` argument.
-        node_list = self.get_node_list(nodes)
-
-        # Check that the front-end is valid
-        if node_list[0].ancestor(GOInvokeSchedule):
-            raise NotImplementedError(
-                "OpenACC kernels regions are not currently supported for "
-                "GOcean InvokeSchedules")
-        super().validate(node_list, options)
-
-        # The regex we use to determine whether a character declaration is
-        # of assumed size ('LEN=*' or '*(*)').
-        assumed_size = re.compile(r"\(\s*len\s*=\s*\*\s*\)|\*\s*\(\s*\*\s*\)")
-
-        # Check that there are no assumed-size character variables as these
-        # causes an Internal Compiler Error with NVHPC.
-        for node in node_list:
-            for lit in node.walk(Literal):
-                if lit.datatype.intrinsic == ScalarType.Intrinsic.CHARACTER:
-                    # We've found a character literal so we go up to the
-                    # ancestor statement and then check the types of all
-                    # symbols that are referenced by it.
-                    stmt = lit.ancestor(Statement)
-                    for ref in stmt.walk(Reference):
-                        if not ref.symbol.is_argument:
-                            # Only arguments can be of assumed length.
-                            continue
-                        # We only need to check the datatype of the underlying
-                        # Symbol.
-                        dtype = ref.symbol.datatype
-                        # Currently the fparser2 frontend does not support any
-                        # type of LEN= specification on a character variable so
-                        # we resort to a regex to check whether it is assumed-
-                        # size.
-                        if isinstance(dtype, UnsupportedFortranType):
-                            type_txt = dtype.type_text.lower()
-                            if (type_txt.startswith("character") and
-                                    assumed_size.search(type_txt)):
-                                raise TransformationError(
-                                    f"Assumed-size character variables cannot "
-                                    f"be enclosed in an OpenACC region but "
-                                    f"found '{stmt.debug_string()}'")
-            # Check that any Intrinsics are supported on the device.
-            for icall in node.walk(IntrinsicCall):
-                if not icall.is_available_on_device():
-                    raise TransformationError(
-                        f"Cannot include intrinsic '{icall.debug_string()}' in"
-                        f" an OpenACC region because it is not available on "
-                        f"GPU.")
-
-        # Check that we have at least one loop or array range within
-        # the proposed region unless this has been disabled.
-        if options and options.get("disable_loop_check", False):
-            return
-
-        for node in node_list:
-            if (any(assign for assign in node.walk(Assignment)
-                    if assign.is_array_assignment) or node.walk(Loop)):
-                break
-        else:
-            # Branch executed if loop does not exit with a break
-            raise TransformationError(
-                "A kernels transformation must enclose at least one loop or "
-                "array range but none were found.")
 
 
 class ACCDataTrans(RegionTrans):
